mini-frontend/requirements.txt
flet>=0.24.0
Pillow>=10.0.0
httpx>=0.27.0
